+++ src/lib.rs
@@ -2,7 +2,10 @@
 
 #![no_std]
 
+pub use accessor;
 pub use error::Error;
+pub use extended_capabilities::ExtendedCapability;
+pub use registers::Registers;
 
 pub mod error;
 pub mod extended_capabilities;